Add form_driver_w
Add wide char version of form_driver with a version of Data_Entry_w as
well, takes the type argument as retrieved from get_wch() as input as
well. The type is used for determining whether the char should be
interpreted as a request or a char that should be entered into the
field.

A test program in test/form_driver_w.c is included.